fix: satisfy mypy --strict across the branch
CI type-checks with mypy --strict and none of these commits had ever been
pushed, so 24 errors accumulated unseen. Mostly annotations that were left
off: untyped test helpers, dict literals inferred too narrowly to hold the
values assigned later, and numpy expressions returning Any through a typed
signature. Two need a word:

The generation-health stub tokenizer implements only the encode/decode
surface the instrument uses, so it is cast rather than widened -- demanding a
real PreTrainedTokenizerBase would put a model's vocabulary between the test
and the behaviour under test.

PackedBlockDataset now binds the reshape to a separate name: rebinding the
memmap to its reshaped view told the checker the variable was two types, and
the reshaped array is a different thing from the flat mapping anyway.

+++ scripts/instrument_ledger.py
@@ -45,7 +45,7 @@ def _correlation(left: list[float], right: list[float]) -> float:
     variance_right = sum((b - mean_right) ** 2 for b in right)
     if variance_left <= 0 or variance_right <= 0:
         return 0.0
-    return covariance / (variance_left * variance_right) ** 0.5
+    return float(covariance / (variance_left * variance_right) ** 0.5)
 
 
 def _sample_corpus(path: Path, limit: int) -> list[str]:
